backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyyaml
alembic
requests
